Linux Fixes
Conditioned the import of two windows only modules, and conditioned their single line usage to only run on the win32 sys platform. I do not believe the usage needs to be changed to have the rest of the function work on linux, but even if it does at the very least squirrel.py is able to be started now.