Ability to setup cache for public links
When a zconfig/remote_cache_location.txt is setup like the remote_cache_location_example.txt pointing to a private drive token and existing location it will be used as caché by copying public links transforming them to auth links.
This avoids errors when reading public links overquota.
Note: The link is copied by gaining ownership, it's not a symlink. This means it takes drive space and daily quota.