Edit server config to make it fixed at 9001 and console visible
So people get it's to open the interface on any browser in the selected host.
Also doesn0t make sense to rotate port with invisible console.
Select hidden console to have an invisible solution at
http:/localhost:9001/main.html
http:/localhost:9001/NSCB.html
NSB.html has the button erased is meant to use with the libraries for remote server.
A remote server can be redirected from a linux machine, for example that was tested to work.